feat(workspace): 「开任务」弹的就是项目主页那个 composer；⌘N / ⌘T / ⌘W；新建项目就地弹框
- 项目主页 composer 抽成 sessions/TaskComposer（状态、提交流程、粘贴图片、话筒都在里面），项目主页和
  弹窗各挂一份；项目行「+」和 ⌘N 弹它，目录预设成当前项目，默认新建 worktree
- 树头「+」就地弹 NewProjectModal，不再先跳到项目列表页再弹（建完跳新项目主页）
- ⌘T 在当前 worktree 里派生终端；⌘W 收起当前标签（文件标签或会话标签）
